[PACK] - Rename cloud SDK package name when importing (#344)
