+++ pasqal_cloud/batch.py
@@ -2,10 +2,10 @@ import time
 from dataclasses import dataclass, field
 from typing import Any, Dict, Optional, Type, Union
 
-from sdk.client import Client
-from sdk.job import Job
-from sdk.device.configuration import BaseConfig, EmuTNConfig, EmuFreeConfig
-from sdk.device import EmulatorType
+from pasqal_cloud.client import Client
+from pasqal_cloud.job import Job
+from pasqal_cloud.device.configuration import BaseConfig, EmuTNConfig, EmuFreeConfig
+from pasqal_cloud.device import EmulatorType
 
 RESULT_POLLING_INTERVAL = 2  # seconds
