+++ pasqal_cloud/device/configuration/base_config.py
