+++ pasqal_cloud/device/configuration/emu_tn.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 
-from sdk.device.configuration.base_config import BaseConfig, InvalidConfiguration
+from pasqal_cloud.device.configuration.base_config import BaseConfig, InvalidConfiguration
 
 DT_VALUE_NOT_VALID = "dt must be larger than 0. Not {}."
 PRECISION_NOT_VALID = "Precision {} not valid. Must be one of 'low', 'normal', 'high'"
