+++ pasqal_cloud/utils/strenum.py
