+++ sdk/README.md
@@ -1,0 +1,5 @@
+# cloud-sdk is now pasqal-cloud
+
+This package has been renamed. Use `pip install pasqal-cloud` instead.
+
+New package: https://pypi.org/project/pasqal-cloud/
