+++ sdk/utils/__init__.py
@@ -1,0 +1,6 @@
+import warnings
+warnings.warn("cloud-sdk package is deprecated, please use pasqal_cloud instead: `pip install pasqal-cloud`",
+              DeprecationWarning)
+
+from pasqal_cloud.utils.strenum import *
+from pasqal_cloud.utils.jsend import *
